backend: auth: Relocate OIDC token refresh middleware and cookie helpers
